+++ src/test/java/com/ivanbasic/learnspring/GreetingController_AnalyzedAgain_ITest.java
@@ -8,12 +8,14 @@ import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
+import org.springframework.test.context.TestPropertySource;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class GreetingControllerAnalyzedAgainITest {
+@TestPropertySource("classpath:application-test.properties")
+public class GreetingController_AnalyzedAgain_ITest {
 
     @Autowired
     private TestRestTemplate template;