+++ src/test/java/com/ivanbasic/learnspring/GreetingController_AnalyzedAgain_Test.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.ResultActions;
@@ -20,7 +21,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @SpringBootTest
 @AutoConfigureMockMvc
-public class GreetingControllerAnalyzedAgainTest {
+@TestPropertySource("classpath:application-test.properties")
+public class GreetingController_AnalyzedAgain_Test {
 
     @Autowired
     private MockMvc mvc;