+++ src/test/resources/data-custom.sql
@@ -1,0 +1,2 @@
+SELECT 1;
+INSERT INTO customer (id,first_name,last_name)  VALUES (1002, 'first', 'last');