+++ src/test/resources/data.sql
@@ -1,0 +1,3 @@
+SELECT 1;
+DELETE FROM customer;
+INSERT INTO customer (id,first_name,last_name)  VALUES (1001, 'first', 'last');